proc: fix the display error of the processes field in /proc/stat

+++ include/linux/sched/stat.h
@@ -15,7 +15,7 @@
 DECLARE_PER_CPU(unsigned long, total_forks);
 extern int nr_threads;
 DECLARE_PER_CPU(unsigned long, process_counts);
-extern int nr_forks(void);
+extern unsigned long nr_forks(void);
 extern int nr_processes(void);
 extern unsigned long nr_running(void);
 extern unsigned long nr_running_cpu(int cpu);
